requirements.txt
altgraph==0.16.1
certifi==2018.11.29
chardet==3.0.4
future==0.17.1
idna==2.8
macholib==1.11
pefile==2018.8.8
PyInstaller==3.4
PyQt5==5.11.3
PyQt5-sip==4.19.13
pywin32-ctypes==0.2.0
PyYAML==3.13
requests==2.21.0
urllib3==1.24.1
